requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.7.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tldextract==3.3.1
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
zope.event==4.5.0
zope.interface==5.4.0
